% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEDF_from_ping_data.R
\name{DEDF_from_ping_data}
\alias{DEDF_from_ping_data}
\title{Create a DEDF from ping- and receiver level detection data}
\usage{
DEDF_from_ping_data(
  dat_in,
  max_delay = lubridate::hours(1),
  min_detect = 2,
  start_time = NULL,
  end_time = NULL
)
}
\arguments{
\item{dat_in}{dataframe with ping-level and receiver level detection, corresponds to  \code{det_pings} output from \code{extract_single_study_ping_data()}}

\item{max_delay}{maximum delay for defining separate detection events}

\item{min_detect}{minimum number of individual instantaneous detection (pings) for defining an 'event'}

\item{start_time}{earliest allowable detection time}

\item{end_time}{latest allowable detection time}
}
\value{
detection event data frame (DEDF)
}
\description{
Reduce ping-level detection data to first and last time at general location event
}
