% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_single_study_ping_data.R
\name{extract_single_study_ping_data}
\alias{extract_single_study_ping_data}
\title{Extract example data set from db using studyID}
\usage{
extract_single_study_ping_data(study, db, start_time = NULL, end_time = NULL)
}
\arguments{
\item{study}{corresponds to 'studyID' variable in 'studies' table of sqlite3 database}

\item{db}{location of sqllite database created for 'predator_filter' gitlab}

\item{start_time}{earliest allowable detection time}

\item{end_time}{latest allowable detection time}
}
\value{
named list of table , including ping-level and receiver level detection \code{det_pings},
deployment ids matched with general locations \code{dep_gl_tb}=gen_loc_sub,tag release information \code{rls}, and release location info \code{rel_loc}
}
\description{
Extract list of tables containing information relevant to each studyID in the sqlite3 database created for 'predator_filter' gitlab
}
