% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ann_HORbar_WYT_data.R
\name{get_ann_HORbar_WYT_data}
\alias{get_ann_HORbar_WYT_data}
\title{Title}
\usage{
get_ann_HORbar_WYT_data(dt_rng_in = c("2011-01-01", "2024-12-31"))
}
\arguments{
\item{dt_rng_in}{character string with two dates}
}
\value{
annual summary of water year type and HOR barrier conditions
}
\description{
Title
}
