% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CNRA_dayflow_by_year.R
\name{get_CNRA_dayflow_by_year}
\alias{get_CNRA_dayflow_by_year}
\title{generate URL for downloading .csv files from CNRA dayflow}
\usage{
get_CNRA_dayflow_by_year(yr_in)
}
\arguments{
\item{yr_in}{single year in}
}
\value{
URL for downloading csvs
}
\description{
based on https://data.cnra.ca.gov/dataset/dayflow/resource/6a7cb172-fb16-480d-9f4f-0322548fee83?view_id=5017ca69-a29c-43f4-9007-20ef0e1bdc54
}
