% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_plot.R
\name{get_env_plot}
\alias{get_env_plot}
\title{Title}
\usage{
get_env_plot(
  env_data_long_in,
  DOY_rng = NULL,
  highlight_years = c(2011:2016),
  yr_high_lb = "Model_years",
  var_fact_levels = c("VNS", "OUT", "CVP", "SWP", "ORB", "MID", "OMT", "CLC", "MSD")
)
}
\arguments{
\item{env_data_long_in}{long-form version of daily environmental covariates}

\item{DOY_rng}{Day of year range}

\item{highlight_years}{years to highlight; detault to mdoel years}

\item{yr_high_lb}{factor variable for highlight years}

\item{var_fact_levels}{factor variable ordering}
}
\value{
ggplot of relevant environmental variables
}
\description{
Title
}
