% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HOR_barrier_data.R
\name{get_HOR_barrier_data}
\alias{get_HOR_barrier_data}
\title{Title}
\usage{
get_HOR_barrier_data(dt_rng = c("2011-01-01", "2016-12-31"), last_year = NULL)
}
\arguments{
\item{dt_rng}{date range}

\item{last_year}{final year where daily barrier TRUE/FALSE values will be reported}
}
\value{
head of old river (HOR) barrier status
}
\description{
Access the Head of Old River (HOR) barrier installation (start and end times), breach, and removal (start and end times)
PDF doc with related info is here:
https://water.ca.gov/-/media/DWR-Website/Web-Pages/Programs/State-Water-Project/Operations-And-Maintenance/Files/Bay-Delta/South-Delta-Temporary-Barriers-Project/History/2023SpringHeadOfOldRiverSch_ADA.pdf
The actual data table is taken from the USGS gitlab 'predator_filter' package files here: data/auxiliary_data/HORB_physical_schedule_1992_2021.csv

Spring HOR barrier not installed after 2019 b/c of high flow in 2019 and because after 2020 the barrier was not required in the NMFS BiOp on Long-term Operations of the CVP and SWP (see 'xx' comment in PDF).
}
