% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOR_CHP_comp.R
\name{HOR_CHP_comp}
\alias{HOR_CHP_comp}
\title{Title}
\usage{
HOR_CHP_comp(
  RData_pth_in = "data/HOR_CHP_mod_dat_ls.RData",
  z_scale_vars = TRUE
)
}
\arguments{
\item{RData_pth_in}{path of model data}

\item{z_scale_vars}{logical, scale and center variables or not}
}
\value{
named list with vital
}
\description{
Title
}
