% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_daily_vals.R
\name{impute_daily_vals}
\alias{impute_daily_vals}
\title{Title}
\usage{
impute_daily_vals(DF_w_in, col_inspect = c("ORB", "MID", "OMT", "MSD", "CLC"))
}
\arguments{
\item{DF_w_in}{daily env data}

\item{col_inspect}{columns to check for NAs}
}
\value{
imputed wide-format dataframe that was input, but with NA valeus replaced with 3-day, 2-sided moving average
}
\description{
Title
}
