% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data_cols.R
\name{scale_data_cols}
\alias{scale_data_cols}
\title{Title}
\usage{
scale_data_cols(
  scale_cols = c("flength", "Tmsd.hor.7dadm", "log.VNS.hor.5", "SWP.hor.5", "CVP.hor.5",
    "Qomt.hor.1net", "Tclc.hor.3"),
  repl_tab_in,
  x.df_SUB_in
)
}
\arguments{
\item{scale_cols}{names of columns to scale and center}

\item{repl_tab_in}{dataframe with colums "pattrn" and "rplace" for dealing with variables with special characters: ":", "()" etcs}

\item{x.df_SUB_in}{full data set}
}
\value{
dataframe with rescaled columns
}
\description{
Scale and center a set of variables (i.e., (x-mean(x))/sd(x))
}
