% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_MA_data.R
\name{sub_MA_data}
\alias{sub_MA_data}
\title{NA substitution using moving average}
\usage{
sub_MA_data(
  DF_w_in,
  col_conv = c("CLC", "VNS", "SWP", "CVP", "MSD"),
  MA_window = c(3, 5, 5, 5, 7),
  fillNAs = TRUE
)
}
\arguments{
\item{DF_w_in}{dataframe with columns to be replaced by moving averages}

\item{col_conv}{column names for conversion}

\item{MA_window}{moving average window size (days)}

\item{fillNAs}{logical, indicating whether NAs at start and end are replaced with estimates from 3-day MA. Defaults to TRUE}
}
\value{
wide-format dataframe that was input, but with some values replaced with various moving averages
}
\description{
Substitute variable specific moving averages for daily values
}
