% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCJ_CHP_comp.R
\name{TCJ_CHP_comp}
\alias{TCJ_CHP_comp}
\title{Title}
\usage{
TCJ_CHP_comp(sdat.det.in, lvec.tcj_ls_in, x.df, z_scale_vars = TRUE)
}
\arguments{
\item{z_scale_vars}{logical, scale and center variables or not}

\item{RData_pth_in}{path of model data}
}
\value{
named list with vital
}
\description{
Title
}
