% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_OMT.R
\name{calc_OMT}
\alias{calc_OMT}
\title{Impute missing values for OMT}
\usage{
calc_OMT(USGS_dat_ls_in, raw = FALSE, wide = F)
}
\arguments{
\item{USGS_dat_ls_in}{data obtained from USGS}

\item{raw}{if TRUE, also prind ORB and MID values}

\item{wide}{if TRUE, output wide format}
}
\value{
wide- or long-form dataframe with OMT value computed (i.e., ORB + MID)
}
\description{
Impute missing values for OMT
}
