% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEDF_from_ping_data.R
\name{DEDF_from_ping_data}
\alias{DEDF_from_ping_data}
\title{Title}
\usage{
DEDF_from_ping_data(
  dat_in,
  max_delay = hours(1),
  min_detect = 2,
  start_time = NULL,
  end_time = NULL
)
}
\arguments{
\item{dat_in}{}

\item{max_delay}{}

\item{min_detect}{}

\item{start_time}{}

\item{end_time}{}
}
\description{
Title
}
