% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_comp.R
\name{env_comp}
\alias{env_comp}
\title{Compile environmental data for CVPAS-steelhead inputs}
\usage{
env_comp(
  url = "https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST",
  basin = "SJ",
  dt_rng = c("2011-01-01", "2016-12-31"),
  DOY_rng = 1:250,
  impute_OMT = TRUE,
  output = "wide"
)
}
\arguments{
\item{url}{defaults to CDEC url of water year types (WYT) 'WSIHIST'}

\item{basin}{basin for WYT lookup; either 'SJ' or 'SAC'}

\item{dt_rng}{character string of length 2 with default date format}

\item{DOY_rng}{character string}

\item{impute_OMT}{if TRUE, fill in missing daily values for OMT using linear imputation ('simputation' package)}

\item{output}{type of data table output to return, "long", "wide" or "list"; defaults to "wide".}
}
\value{
dataframe or list with combined, daily-scale environmental variabltes
}
\description{
Compile environmental data for CVPAS-steelhead inputs
}
\seealso{
get_CDEC_data()
}
