% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HOR_barrier_data.R
\name{get_barrier_strt_stop}
\alias{get_barrier_strt_stop}
\title{Title}
\usage{
get_barrier_strt_stop()
}
\value{
table of years between 1992 and 2021 where HOR barrier was in place
}
\description{
Hard-coding of spring HOR barrier
}
