% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CNRA_dayflow_by_year.R
\name{get_dayflow_data}
\alias{get_dayflow_data}
\title{Title}
\usage{
get_dayflow_data(dt_rng = c("2011-01-01", "2016-12-31"))
}
\arguments{
\item{dt_rng}{date range}
}
\value{
dataframe with dayflow measures from CNRA
}
\description{
Accessing Dayflow datatables in chunks broken up by year
Dayflow summary:
https://water.ca.gov/Programs/Integrated-Science-and-Engineering/Compliance-Monitoring-And-Assessment/Dayflow-Data
}
